Allow 3705 devices to be used when statically linking
Adding an entry for 3705 devices to hdlmain.c allows these devices to be
used when Hercules is statically linked.  However, this results in a linker
warning that symbol get_inet_socket() is multiply defined (in comm3705.c
and in console.c).  However, get_inet_socket() is not called in comm3705.c
and it is also surrounded by #if 1 / #endif so it seems reasonable to
change #if 1 to #if 0 in comm3705.c